Narrow.cpp: explicit uint64_t cast so the pass builds on Linux
(uint64_t)-1 & ((1ULL << To) - 1) casts only the -1; the & result takes
the type of 1ULL, unsigned long long. On macOS uint64_t IS unsigned long
long, so z3::context::bv_val(uint64_t, unsigned) matched exactly. On
Linux uint64_t is unsigned long, and Z3 4.8.12's z3++.h has no unsigned
long long overload, so the call was ambiguous among four candidates and
the c220g2 build failed. Wrap the whole expression in (uint64_t)(...), as
line 98 of the same file already did. Same value, so no verdict can
change: run_tests 27/12 and run_mv_tests 9/9 unchanged on the Mac.

Found on the first Linux build of the September machinery (Narrow.cpp
postdates the August server campaign).

+++ OraclePass/Narrow.cpp
@@ -211,10 +211,10 @@ z3::expr Narrower::go(const z3::expr &E) {
         if (WasFrom) flag((A[0] < Ctx.bv_val(0, To)) || (A[1] < Ctx.bv_val(0, To)), "urem neg", E);
         R = z3::urem(A[0], A[1]); break;
     case Z3_OP_BSDIV:
-        if (WasFrom) flag((A[0] == IntMin) && (A[1] == Ctx.bv_val((uint64_t)-1 & ((1ULL << To) - 1), To)), "sdiv ovf", E);
+        if (WasFrom) flag((A[0] == IntMin) && (A[1] == Ctx.bv_val((uint64_t)((uint64_t)-1 & ((1ULL << To) - 1)), To)), "sdiv ovf", E);
         R = A[0] / A[1]; break;
     case Z3_OP_BSREM:
-        if (WasFrom) flag((A[0] == IntMin) && (A[1] == Ctx.bv_val((uint64_t)-1 & ((1ULL << To) - 1), To)), "srem ovf", E);
+        if (WasFrom) flag((A[0] == IntMin) && (A[1] == Ctx.bv_val((uint64_t)((uint64_t)-1 & ((1ULL << To) - 1)), To)), "srem ovf", E);
         R = z3::srem(A[0], A[1]); break;
     // ---- comparisons (exact on fitting values)
     case Z3_OP_ULEQ: R = z3::ule(A[0], A[1]); break;
